Refactored code so that resource schema and read functions appear in the resource file.
