+++ pkg/artifactory/provider/provider.go
@@ -168,7 +168,7 @@ func Provider() *schema.Provider {
 			map[string]*schema.Resource{
 				"artifactory_file":     datasource.ArtifactoryFile(),
 				"artifactory_fileinfo": datasource.ArtifactoryFileInfo(),
-				"artifactory_group":    datasource.ArtifactoryGroupSchema(),
+				"artifactory_group":    datasource.ArtifactoryGroup(),
 			},
 		),
 	}
